Set summary for auto-freeze tags.
Using this code to set the summary onPreferenceChanged: https://stackoverflow.com/questions/41775332/showing-selected-values-in-summary-from-multiselectlistpreference
Thanks for that!!

Haven't (yet) managed to keep track of renamed or deleted tabs. Because of
this, add a failsafe to HailData.autoFreezeTags, which makes sure no
autoFreezeTags are used if there isn't an actual tag with that tagId.