Make AutoFreezeService a little more persistent (#55)
* Make AutoFreezeService a little more persistent.

Don't stop service from ScreenOffReceiver directly but from AutoFreezeWorker
once the job is done.

In some edge cases, the AutoFreezeService might noz run. E.g. ff the screen
is turned on quickly after turning off, the AutoFreezeWorker won't freeze
the apps precisely because the screen is on. Because the AutoFreezeService
isn't running anymore, it won't try to freeze the apps on the next switching
off of the screen.

With this commit this, the service won't stop until it has done its
job during a turning off of the screen.

* Stop AutoFreezeService when switched off in settings.

There is no need to check autoFreezeAfterLock anymore while running
AutoFreezeWorker because service is already being disabled when
the preference changes.